server/requirements.txt
aioredis==1.3.1
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
constantly==15.1.0
cryptography==3.4.6
daphne==3.0.1
Django==3.2
django-computedfields==0.1.3
django-cors-headers==3.7.0
django-dbsettings==1.1.0
django-environ==0.4.5
django-mathfilters==1.0.0
djangorestframework==3.12.4
hiredis==2.0.0
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
msgpack==1.0.2
Pillow==8.1.0
psycopg2==2.8.6
ptvsd==4.3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2020.5
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
zope.interface==5.2.0
